+++ src/index.js
@@ -4,3 +4,4 @@ export { default as PersianDateTimePicker } from './PersianDateTimePicker.jsx';
 export { default as PersianDateRangePicker } from './PersianDateRangePicker.jsx';
 export { default as PersianCalendar } from './PersianCalendar.jsx';
 export { default as PersianTimePicker } from './PersianTimePicker.jsx';
+export { default as PersianTimeline } from './PersianTimeline.jsx';